+++ src/main/java/org/ohdsi/standardized_analysis_api/Cohort.java
@@ -1,4 +1,4 @@
-package org.ohdsi.standardized_analysis_api.cohortcharacterization.design;
+package org.ohdsi.standardized_analysis_api;
 
 import com.fasterxml.jackson.annotation.JsonGetter;
 import org.ohdsi.circe.cohortdefinition.CohortExpression;
