+++ src/main/java/org/ohdsi/standardized_analysis_api/CohortMetadata.java
@@ -1,4 +1,4 @@
-package org.ohdsi.standardized_analysis_api.cohortcharacterization.design;
+package org.ohdsi.standardized_analysis_api;
 
 import com.fasterxml.jackson.annotation.JsonGetter;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
